Split lib/token.py into a token/ package
The 3115-line privacyidea/lib/token.py is broken up into a package with
submodules grouped by responsibility along the (cycle-free) call graph:

  query        token lookup/retrieval (leaf dependency)
  attributes   the set_* / tokeninfo / enable/revoke/reset family
  auth         the check_* validation family
  lifecycle    create/import/assign/copy/delete
  otp          OTP retrieval and reverse lookup
  tokengroups  token-group assignment
  importexport bulk import/export
  misc         policy definitions and enrollment-URL helpers
  const        shared constants

__init__.py re-exports every public name, so existing
"from privacyidea.lib.token import X" call sites keep working unchanged.
get_token_types is re-exported explicitly because it was previously
exposed transitively through this module.
