+++ privacyidea/lib/token/const.py
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: 2026 NetKnights GmbH <https://netknights.it>
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""Shared constants for the token library package."""
+
+ENCODING = "utf-8"
+
+# Configuration to generate a complete random serial
+PI_TOKEN_SERIAL_RANDOM = "PI_TOKEN_SERIAL_RANDOM"  # nosec B105
+
+B32_ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567'
